End first refactoring
