+++ src/controller/commandController.ts
@@ -1,13 +1,14 @@
 import * as vscode from 'vscode';
-// import { refreshUnityProject, findMetaReference } from '../loader';
+import path = require('path');
+import { refreshUnityProject } from './unityProjectController';
 
 export function updateStatus<T>(name: string, value: T) {
     vscode.commands.executeCommand('setContext', name, value);
 }
 
-export function initialize(context: vscode.ExtensionContext) {
-    // registerCommand(context, 'clover.findMetaReference', () => findMetaReference());
-	// registerCommand(context, 'clover.refreshUnityProject', () => refreshUnityProject());
+export function initialize(context: vscode.ExtensionContext, workspacePath: string) {
+    const assetPath = path.join(workspacePath, 'Assets');
+	registerCommand(context, 'clover.refreshUnityProject', () => refreshUnityProject(assetPath));
 	registerCommand(context, 'clover.noReferenceMessage', () => vscode.window.showInformationMessage("No reference found"));
 }
