--- src/metaData.ts
@@ -1,7 +1,0 @@
-export class MetaData {
-    path: string;
-
-    constructor(path: string) {
-        this.path = path;
-    }
-}
