+++ src/parser/guidConnector.ts
