+++ src/provider/mainViewProvider.ts
@@ -1,20 +1,12 @@
 import * as vscode from 'vscode';
-import * as parser from '../guidParser';
 import fs = require('fs');
 import path = require('path');
 
 export class MainViewProvider implements vscode.WebviewViewProvider {
     private _extensionUri: vscode.Uri;
-    // private _projectName: string;
-    // private _projectVersion: string;
 
-    constructor(extensionUri: vscode.Uri, projectPath: string) {
+    constructor(extensionUri: vscode.Uri) {
         this._extensionUri = extensionUri;
-        var projectSettingsPath = path.join(projectPath, "ProjectSettings", "ProjectSettings.asset");
-        var projectVersionPath = path.join(projectPath, "ProjectSettings", "ProjectVersion.txt");
-        // this._projectName = parser.getProjectName(fs.readFileSync(projectSettingsPath, { encoding: 'utf8' }));
-        // this._projectVersion = parser.getProjectVersion(fs.readFileSync(projectVersionPath, { encoding: 'utf8' }));
-
     }
 
     resolveWebviewView(webviewView: vscode.WebviewView, context: vscode.WebviewViewResolveContext<unknown>, token: vscode.CancellationToken) {
@@ -25,7 +17,7 @@ export class MainViewProvider implements vscode.WebviewViewProvider {
         return `
             <html>
             <body>
-            
+                <h2>Clover for unity</h1>
             </body>
             </html>`;
     }
